requirements.txt
alembic==0.7.5.post2
Flask==0.10.1
Flask-Bootstrap==3.3.2.1
flask-jsontools==0.1.1.post0
Flask-Migrate==1.4.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
gnureadline==6.3.3
gviz-api.py==1.8.2
ipdb==0.8.1
ipython==3.1.0
itsdangerous==0.24
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
SQLAlchemy==1.0.2
Werkzeug==0.10.4
WTForms==2.0.2
